# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Community Assistent System",
        "title": "OverflowStack",
        "contact": {}
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/discuss/comments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "You can comment a discussion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Comment discussion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of discussion",
                        "name": "discussionID",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Your comment",
                        "name": "content",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discuss/discussions": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "You can post new discussion",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Create New Discussion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Title of discussion",
                        "name": "title",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Describe your problem here",
                        "name": "content",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discuss/vote": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit a vote with either \"like\" or \"dislike\"",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Submit a vote",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of discussion or comment",
                        "name": "ElementId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The type of vote. Can be either 'like' or 'dislike'.",
                        "name": "vote",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/discussions": {
            "get": {
                "description": "Get all discussions on site",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Get all discussions",
                "responses": {}
            }
        },
        "/getdiscussion": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get full display of discussion with comments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Get full discussion",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Id of discussion",
                        "name": "discussion_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/search": {
            "get": {
                "description": "Get all discussions on site",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "discussions"
                ],
                "summary": "Get all discussions",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "discussionName",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/users/login": {
            "post": {
                "description": "create account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignIn",
                "parameters": [
                    {
                        "type": "string",
                        "description": "your email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "your password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/users/register": {
            "post": {
                "description": "create account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "SignUp",
                "parameters": [
                    {
                        "type": "string",
                        "description": "your username",
                        "name": "username",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "your password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "your email",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}